refactor: mirror core and integration test placement
